requirements.txt
Box2D==2.3.2
Box2D-kengz==2.3.3
box2d-py==2.3.5
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
Farama-Notifications==0.0.4
fonttools==4.39.3
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
jax-jumpy==1.0.0
kiwisolver==1.4.4
matplotlib==3.7.1
neat-python==0.92
numpy==1.24.2
packaging==23.1
Pillow==9.5.0
pygame==2.1.3
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
swig==4.1.1
typing_extensions==4.5.0
